verdin-imx8mp-5inch: tweak lvds PLL frequency
Due to limitations in the nxp kernel, the frequency of LVDS displays was
fixed for 1 channel displays to 74MHz. This is not compatible with our
5-inch display which expects 25MHz. Remove this patch.

We revert the PLL parent frequency to its value in the 5.15 kernel which
was compatible with all 3 OPTO logic displays (5, 7 and 10.1 inches).

Source: - https://developer.toradex.com/linux-bsp/application-development/multimedia/display-output-resolution-and-timings-linux/?utm_source=chatgpt.com#verdin-imx8m-plus-lvds-known-issues
Change-Id: I30638a76447a6cc6781001f040e28bdd2c48fc9c